requirements.txt
numpy>=1.19.1
pandas>=1.1.1
pillow~=7.2.0
torchvision~=0.7.0
matplotlib~=3.3.1
opencv>=4.4.0
pandas>=1.1.1
scipy>=1.5.2